added breeze and finished up crash course

+++ app/Http/Controllers/NoteController.php
@@ -12,7 +12,10 @@ class NoteController extends Controller
      */
     public function index()
     {
-        $notes = Note::query()->orderBy("created_at", 'desc')->paginate();
+        $notes = Note::query()
+            ->where('user_id', request()->user()->id)
+            ->orderBy("created_at", 'desc')
+            ->paginate();
         return view('note.index', ['notes' => $notes]);
     }
 
@@ -33,7 +36,7 @@ class NoteController extends Controller
         $data = $request->validate(
             ['note' => ['required', 'string']]
         );
-        $data['user_id'] = 1;
+        $data['user_id'] = $request->user()->id;
         $note = Note::create($data);
 
         return to_route('note.show', $note)->with('message', 'Note was created');
@@ -44,6 +47,9 @@ class NoteController extends Controller
      */
     public function show(Note $note)
     {
+        if ($note->user_id !== request()->user()->id) {
+            return abort(403);
+        }
         return view('note.show', ['note' => $note]);
 
     }
@@ -53,6 +59,9 @@ class NoteController extends Controller
      */
     public function edit(Note $note)
     {
+        if ($note->user_id !== request()->user()->id) {
+            return abort(403);
+        }
         return view('note.edit', ['note' => $note]);
 
     }
@@ -62,6 +71,9 @@ class NoteController extends Controller
      */
     public function update(Request $request, Note $note)
     {
+        if ($note->user_id !== request()->user()->id) {
+            return abort(403);
+        }
         $data = $request->validate(
             ['note' => ['required', 'string']]
         );
@@ -76,6 +88,9 @@ class NoteController extends Controller
      */
     public function destroy(Note $note)
     {
+        if ($note->user_id !== request()->user()->id) {
+            return abort(403);
+        }
         $note->delete();
 
         return to_route('note.index', $note)->with('message', 'Note was deleted');
